@@ -5,9 +5,10 @@
   "main": "lib/index.js",
   "scripts": {
     "build": "npm run build:lib && npm run build:ssr",
-    "build:lib": "rimraf lib && babel src/lib -d lib/",
-    "build:ssr": "rimraf ssr && babel src/ssr -d ssr/",
-    "lint": "eslint ./",
+    "build:lib": "rimraf lib && cross-env NODE_ENV=production babel src/lib -d lib/ --source-maps inline",
+    "build:ssr": "rimraf ssr && cross-env NODE_ENV=production babel src/ssr -d ssr/",
+    "lint": "eslint ./src ./test",
+    "lint:fix": "eslint --fix ./src ./test",
     "prepublish": "npm run build",
     "test": "nyc ava --verbose",
     "report": "nyc report --reporter=html"
@@ -17,40 +18,43 @@
     "ssr"
   ],
   "peerDependencies": {
-    "react": "*",
-    "rxjs": "^5.0.0"
+    "react": "^15.0.0 || ^16.0.0",
+    "rxjs": "^6.0.0"
   },
   "dependencies": {
     "history": "^4.7.2",
-    "path-to-regexp": "^2.0.0"
+    "path-to-regexp": "^2.2.1"
   },
   "devDependencies": {
-    "ava": "^0.22.0",
-    "babel-cli": "^6.26.0",
-    "babel-eslint": "^7.2.3",
-    "babel-loader": "^7.1.2",
-    "babel-plugin-transform-async-to-generator": "^6.24.1",
-    "babel-plugin-transform-decorators-legacy": "^1.3.4",
-    "babel-plugin-transform-runtime": "^6.23.0",
-    "babel-plugin-unassert": "^2.1.2",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-1": "^6.24.1",
-    "browser-env": "^3.2.0",
-    "coveralls": "^2.13.1",
-    "enzyme": "^2.9.1",
-    "eslint": "^4.5.0",
-    "eslint-plugin-import": "^2.7.0",
-    "eslint-plugin-react": "^7.3.0",
-    "just-extend": "1.1.22",
-    "nyc": "^11.1.0",
-    "react": "^15.6.1",
-    "react-dom": "^15.6.1",
-    "react-test-renderer": "^15.6.1",
-    "rimraf": "^2.6.1",
-    "rxjs": "^5.4.3",
-    "sinon": "^4.0.1",
-    "webpack": "^3.5.5"
+    "@babel/cli": "^7.0.0-rc.1",
+    "@babel/core": "^7.0.0-rc.1",
+    "@babel/plugin-transform-runtime": "^7.0.0-rc.1",
+    "@babel/preset-env": "^7.0.0-rc.1",
+    "@babel/preset-react": "^7.0.0-rc.1",
+    "@babel/register": "^7.0.0-rc.1",
+    "@babel/runtime": "^7.0.0-rc.1",
+    "ava": "1.0.0-beta.7",
+    "babel-eslint": "^8.2.6",
+    "babel-loader": "^8.0.0-beta.0",
+    "babel-plugin-istanbul": "^4.1.6",
+    "browser-env": "^3.2.5",
+    "coveralls": "^3.0.2",
+    "cross-env": "^5.2.0",
+    "enzyme": "^3.3.0",
+    "enzyme-adapter-react-16": "^1.2.0",
+    "eslint": "^5.3.0",
+    "eslint-config-prettier": "^3.0.1",
+    "eslint-plugin-import": "^2.13.0",
+    "eslint-plugin-prettier": "^2.6.2",
+    "nyc": "^12.0.2",
+    "prettier": "1.14.2",
+    "react": "^16.4.2",
+    "react-dom": "^16.4.2",
+    "react-test-renderer": "^16.4.2",
+    "rimraf": "^2.6.2",
+    "rxjs": "^6.2.2",
+    "sinon": "^6.1.4",
+    "webpack": "^4.16.4"
   },
   "repository": {
     "type": "git",
@@ -72,9 +76,22 @@
     "concurrency": 5,
     "powerAssert": true,
     "require": [
-      "babel-register",
+      "@babel/register",
       "./test/helpers/setup-browser-env.js"
     ],
-    "babel": "inherit"
+    "babel": {
+      "testOptions": {
+        "plugins": [
+          "@babel/plugin-transform-runtime"
+        ]
+      }
+    }
+  },
+  "nyc": {
+    "include": [
+      "src/**/*.js"
+    ],
+    "sourceMap": false,
+    "instrument": false
   }
 }
