@@ -1,17 +1,26 @@
-import Request from "./Request";
 import React from "react";
 
-export default class Link extends React.Component {
-    click(e) {
-        e.preventDefault();
-        Request.to(this.props.to);
-    }
+class Link extends React.Component {
+  click(e) {
+    e.preventDefault();
+    this.props.request.to(this.props.to);
+  }
 
-    render() {
-        return (
-            <a href={this.props.to} className={this.props.className} onClick={this.click.bind(this)}>
-                {this.props.children}
-            </a>
-        );
-    }
+  render() {
+    return (
+      <a
+        href={this.props.to}
+        className={this.props.className}
+        onClick={this.click.bind(this)}
+      >
+        {this.props.children}
+      </a>
+    );
+  }
+}
+
+export function createLink(request) {
+  return (props) => {
+    return <Link request={request} {...props} />;
+  };
 }
