@@ -1,5 +1,5 @@
-const Route = function () {
-    // Route should not be rendered
+const Route = function() {
+  // Route should not be rendered
 };
 
 export default Route;
