@@ -1,24 +1,17 @@
-import HistoryManager from "./HistoryManager";
+export default class URL {
+  constructor() {
+    this.historyManager = null;
+  }
 
-class URL {
-    constructor() {
-        this.historyManager = null;
-    }
+  setHistoryManager(historyManager) {
+    this.historyManager = historyManager;
+  }
 
-    setHistoryManager(historyManager) {
-        this.historyManager = historyManager;
-    }
+  to(pathname) {
+    return this.historyManager.createHref(pathname);
+  }
 
-    to(pathname) {
-        return this.historyManager.createHref(pathname);
-    }
-
-    name(name, parameters = {}) {
-        return this.historyManager.createHrefByName(name, parameters);
-    }
+  name(name, parameters = {}) {
+    return this.historyManager.createHrefByName(name, parameters);
+  }
 }
-
-const url = new URL();
-url.setHistoryManager(HistoryManager);
-
-export default url;
