@@ -1,19 +1,19 @@
 import createBrowserHistory from "history/createBrowserHistory";
 import createHashHistory from "history/createHashHistory";
 import createMemoryHistory from "history/createMemoryHistory";
-import Link from "./Link";
-import Request from "./Request";
+import { Link, Request, URL, routeMatcher, historyManager } from "./Facade";
 import Route from "./Route";
-import Router from "./Router";
-import URL from "./URL";
+import { createRouter } from "./Router";
+
+const router = createRouter(historyManager, routeMatcher);
 
 export {
-    createBrowserHistory,
-    createHashHistory,
-    createMemoryHistory,
-    Link,
-    Request,
-    Route,
-    Router,
-    URL
+  createBrowserHistory,
+  createHashHistory,
+  createMemoryHistory,
+  Link,
+  Request,
+  Route,
+  router as Router,
+  URL
 };
