@@ -1,7 +1,9 @@
-import Router from "./Router";
 import RouteResolver from "./RouteResolver";
+import { createRouter } from "./Router";
+import { routeMatcher, historyManager } from "../lib/Facade";
 
-export {
-    Router,
-    RouteResolver
-};
+const router = createRouter(historyManager, routeMatcher);
+
+const routeResolver = new RouteResolver(historyManager, routeMatcher);
+
+export { router as Router, routeResolver as RouteResolver };
