@@ -1,90 +1,96 @@
-import Test from "ava";
+import test from "ava";
 import HistoryManager from "../../src/lib/HistoryManager";
 import createMemoryHistory from "history/createMemoryHistory";
 
-Test.serial('push path', async (t) => {
-    const historyCallBack = (pathname) => {
-        t.is(pathname, '/test');
-    };
+test.serial("push path", async (t) => {
+  const historyCallBack = (pathname) => {
+    t.is(pathname, "/test");
+  };
 
-    HistoryManager.init();
-    HistoryManager.initialHistory(createMemoryHistory());
-    HistoryManager.setRequestCallback(historyCallBack);
-    HistoryManager.listen();
-    await HistoryManager.push('/test');
+  const historyManager = new HistoryManager();
+  historyManager.initialHistory(createMemoryHistory());
+  historyManager.setRequestCallback(historyCallBack);
+  historyManager.listen();
+  await historyManager.push("/test");
 
-    const location = HistoryManager.getLocation();
-    t.is(location.pathname, '/test');
+  const location = historyManager.getLocation();
+  t.is(location.pathname, "/test");
 });
 
-Test.serial('push path by name', async (t) => {
-    class RouteMatcherMock {
-        compileByName(name, parameters = {}) {
-            return '/test/1';
-        }
+test.serial("push path by name", async (t) => {
+  class RouteMatcherMock {
+    compileByName(name, parameters = {}) {
+      return "/test/1";
     }
+  }
 
-    const historyCallBack = (pathname) => {
-        t.is(pathname, '/test/1');
-    };
+  const historyCallBack = (pathname) => {
+    t.is(pathname, "/test/1");
+  };
 
-    HistoryManager.init();
-    HistoryManager.initialHistory(createMemoryHistory());
-    HistoryManager.initialRouteMatcher(new RouteMatcherMock());
-    HistoryManager.setRequestCallback(historyCallBack);
-    HistoryManager.listen();
-    await HistoryManager.pushByName('Test', {test: 1});
+  const historyManager = new HistoryManager();
+  historyManager.initialHistory(createMemoryHistory());
+  historyManager.initialRouteMatcher(new RouteMatcherMock());
+  historyManager.setRequestCallback(historyCallBack);
+  historyManager.listen();
+  await historyManager.pushByName("Test", { test: 1 });
 
-    const location = HistoryManager.getLocation();
-    t.is(location.pathname, '/test/1');
+  const location = historyManager.getLocation();
+  t.is(location.pathname, "/test/1");
 });
 
-Test.serial('create path', (t) => {
-    HistoryManager.initialHistory(createMemoryHistory());
-    t.is(HistoryManager.createHref('/create_href'), '/create_href');
+test.serial("create path", (t) => {
+  const historyManager = new HistoryManager();
+  historyManager.initialHistory(createMemoryHistory());
+  t.is(historyManager.createHref("/create_href"), "/create_href");
 });
 
-Test.serial('listen silent', async (t) => {
-    const historyCallBack = (pathname) => {
-        t.fail();
-    };
+test.serial("listen silent", async (t) => {
+  const historyCallBack = (pathname) => {
+    t.fail();
+  };
 
-    const location = {
-        pathname: '/test'
-    };
-    HistoryManager.init();
-    HistoryManager.setRequestCallback(historyCallBack);
-    HistoryManager.changeSilent();
-    HistoryManager.listenCallback(location);
-    t.pass();
+  const location = {
+    pathname: "/test"
+  };
+
+  const historyManager = new HistoryManager();
+  historyManager.setRequestCallback(historyCallBack);
+  historyManager.changeSilent();
+  historyManager.listenCallback(location);
+  t.pass();
 });
 
-Test.serial('listen unsilent', async (t) => {
-    const historyCallBack = (pathname) => {
-        t.is(pathname, '/unsilent');
-    };
+test.serial("listen unsilent", async (t) => {
+  const historyCallBack = (pathname) => {
+    t.is(pathname, "/unsilent");
+  };
 
-    const location = {
-        pathname: '/unsilent'
-    };
-    HistoryManager.init();
-    HistoryManager.setRequestCallback(historyCallBack);
-    HistoryManager.changeSilent();
-    HistoryManager.changeUnsilent();
-    HistoryManager.listenCallback(location);
+  const location = {
+    pathname: "/unsilent"
+  };
+
+  const historyManager = new HistoryManager();
+  historyManager.setRequestCallback(historyCallBack);
+  historyManager.changeSilent();
+  historyManager.changeUnsilent();
+  historyManager.listenCallback(location);
 });
 
-Test.serial('create path by name', (t) => {
-    class RouteMatcherMock {
-        compileByName(name, parameters = {}) {
-            t.is(name, 'create_href_by_name');
-            t.is(parameters.test, 1);
-            return '/create_href_by_name/1';
-        }
+test.serial("create path by name", (t) => {
+  class RouteMatcherMock {
+    compileByName(name, parameters = {}) {
+      t.is(name, "create_href_by_name");
+      t.is(parameters.test, 1);
+      return "/create_href_by_name/1";
     }
+  }
 
-    HistoryManager.init();
-    HistoryManager.initialHistory(createMemoryHistory());
-    HistoryManager.initialRouteMatcher(new RouteMatcherMock());
-    t.is(HistoryManager.createHrefByName('create_href_by_name', {test: 1}), '/create_href_by_name/1');
+  const historyManager = new HistoryManager();
+  historyManager.initialHistory(createMemoryHistory());
+  historyManager.initialRouteMatcher(new RouteMatcherMock());
+  t.is(
+    historyManager.createHrefByName("create_href_by_name", { test: 1 }),
+    "/create_href_by_name/1"
+  );
 });
