@@ -1,95 +1,191 @@
-import React from 'react';
-import test from 'ava';
-import RouteMatcher from '../../src/lib/RouteMatcher';
+import React from "react";
+import test from "ava";
+import RouteMatcher from "../../src/lib/RouteMatcher";
 
-test('Add Route', (t) => {
-    class TestComponent extends React.Component {
-        render() {
-            return (<div>Hello, World</div>);
-        }
+test("Add route", (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
     }
+  }
 
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/Test', TestComponent);
-    t.is(RouteMatcher.routes[0].path, '/Test');
-    t.is(RouteMatcher.routes[0].component, TestComponent);
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/Test/:id', TestComponent, 'test');
-    t.is(RouteMatcher.routes[0].path, '/Test/:id');
-    t.is(RouteMatcher.routes[0].component, TestComponent);
-    t.is(RouteMatcher.nameRoutes['test'], '/Test/:id');
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/Test", TestComponent);
+  t.is(routeMatcher.routes[0].path, "/Test");
+  t.is(routeMatcher.routes[0].component, TestComponent);
 });
 
-test('Fetch renderer', (t) => {
-    class TestComponent extends React.Component {
-        render() {
-            return (<div>Hello, World</div>);
-        }
+test("Add Route with url params", (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
     }
+  }
 
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/int', TestComponent);
-    const renderer1 = RouteMatcher.fetchRenderer('/int').getRenderer();
-    const prams1    = renderer1.getProps().params;
-    t.is(renderer1.getComponent(), TestComponent);
-    t.true(Object.keys(prams1).length === 0 && typeof prams1 === "object");
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/int/:id', TestComponent);
-    const renderer2 = RouteMatcher.fetchRenderer('/int/1').getRenderer();
-    const prams2    = renderer2.getProps().params;
-    t.is(renderer2.getComponent(), TestComponent);
-    t.is(prams2.id, 1);
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/int/:id', TestComponent);
-    const renderer3 = RouteMatcher.fetchRenderer('/int/1/test').getRenderer();
-    t.is(renderer3, null);
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/string/:name', TestComponent);
-    const renderer4 = RouteMatcher.fetchRenderer('/string/name').getRenderer();
-    const prams4    = renderer4.getProps().params;
-    t.is(renderer4.getComponent(), TestComponent);
-    t.is(prams4.name, 'name');
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/string/:name+', TestComponent);
-    const renderer5 = RouteMatcher.fetchRenderer('/string').getRenderer();
-    t.is(renderer5, null);
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/string/:name+', TestComponent);
-    const renderer6 = RouteMatcher.fetchRenderer('/string/name').getRenderer();
-    const prams6    = renderer6.getProps().params;
-    t.is(renderer6.getComponent(), TestComponent);
-    t.is(prams6.name, 'name');
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/Test/:id", TestComponent, "test");
+  t.is(routeMatcher.routes[0].path, "/Test/:id");
+  t.is(routeMatcher.routes[0].component, TestComponent);
+  t.is(routeMatcher.nameRoutes["test"], "/Test/:id");
 });
 
-test('Compile by name', (t) => {
-    class TestComponent extends React.Component {
-        render() {
-            return (<div>Hello, World</div>);
-        }
+test("Fetch renderer on index", async (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
     }
+  }
 
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/Test', TestComponent, 'test');
-    const url = RouteMatcher.compileByName('test');
-    t.is(url, '/Test');
-
-    RouteMatcher.init();
-    RouteMatcher.addRoute('/Test/:id', TestComponent, 'test');
-    const url2 = RouteMatcher.compileByName('test', {id: 1});
-    t.is(url2, '/Test/1');
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/index", TestComponent);
+  const renderer = (await routeMatcher.fetchRenderer("/index")).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), TestComponent);
+  t.true(Object.keys(prams).length === 0 && typeof prams === "object");
 });
 
-test('Error compile by name ', (t) => {
-    const error = t.throws(() => {
-        RouteMatcher.init();
-        RouteMatcher.compileByName('test');
-    }, Error);
+test("Fetch renderer on show", async (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
 
-    t.is(error.message, `Route Name "test" did not match Path.`);
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/show/:id", TestComponent);
+  const renderer = (await routeMatcher.fetchRenderer("/show/1")).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), TestComponent);
+  t.is(prams.id, 1);
+});
+
+test("Fetch renderer with int params", async (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/show/:id", TestComponent);
+  const renderer = (await routeMatcher.fetchRenderer("/show/1")).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), TestComponent);
+  t.is(prams.id, 1);
+});
+
+test("Fetch renderer with string params", async (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/string/:name", TestComponent);
+  const renderer = (await routeMatcher.fetchRenderer(
+    "/string/name"
+  )).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), TestComponent);
+  t.is(prams.name, "name");
+});
+
+test("Fetch renderer on not found", async (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/show/:id", TestComponent);
+  const renderer = (await routeMatcher.fetchRenderer(
+    "/show/1/test"
+  )).getRenderer();
+  t.is(renderer, null);
+});
+
+test("Fetch renderer with __esModule on dynamic imports", async (t) => {
+  class EsModuleComponent extends React.Component {
+    render() {
+      return <div>Hello, esModule.</div>;
+    }
+  }
+
+  function importMock() {
+    return new Promise((resolve) => {
+      resolve({
+        default: EsModuleComponent,
+        __esModule: true
+      });
+    });
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/esmodule", importMock, "esmodule", true);
+  const renderer = (await routeMatcher.fetchRenderer(
+    "/esmodule"
+  )).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), EsModuleComponent);
+  t.true(Object.keys(prams).length === 0 && typeof prams === "object");
+});
+
+test("Fetch renderer without __esModule on dynamic imports", async (t) => {
+  class EsModuleComponent extends React.Component {
+    render() {
+      return <div>Hello, esModule.</div>;
+    }
+  }
+
+  function importMock() {
+    return new Promise((resolve) => {
+      resolve(EsModuleComponent);
+    });
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/unesmodule", importMock, "unesmodule", true);
+  const renderer = (await routeMatcher.fetchRenderer(
+    "/unesmodule"
+  )).getRenderer();
+  const prams = renderer.getProps().params;
+  t.is(renderer.getComponent(), EsModuleComponent);
+  t.true(Object.keys(prams).length === 0 && typeof prams === "object");
+});
+
+test("Compile by name", (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/Test", TestComponent, "test");
+  const url = routeMatcher.compileByName("test");
+  t.is(url, "/Test");
+});
+
+test("Compile by name with url params", (t) => {
+  class TestComponent extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  const routeMatcher = new RouteMatcher();
+  routeMatcher.addRoute("/Test/:id", TestComponent, "test");
+  const url = routeMatcher.compileByName("test", { id: 1 });
+  t.is(url, "/Test/1");
+});
+
+test("Error compile by name ", (t) => {
+  const error = t.throws(() => {
+    const routeMatcher = new RouteMatcher();
+    routeMatcher.compileByName("test");
+  }, Error);
+
+  t.is(error.message, `Route Name "test" did not match Path.`);
 });
