@@ -1,268 +1,296 @@
 import React from "react";
-import Test from "ava";
-import Route from "../../src/lib/Route";
-import Router from "../../src/lib/Router";
-import Request from "../../src/lib/Request";
+import test from "ava";
 import { mount } from "enzyme";
 import createMemoryHistory from "history/createMemoryHistory";
-import Sinon from "sinon";
+import Route from "../../src/lib/Route";
+import { createRouter } from "../../src/lib/Router";
+import Request from "../../src/lib/Request";
+import RouteMatcher from "../../src/lib/RouteMatcher";
+import HistoryManager from "../../src/lib/HistoryManager";
 
-Test.serial('match single route', async (t) => {
-    // Location Settings
-    const history = createMemoryHistory();
-    history.push('/');
+const asyncFlush = () => new Promise((resolve) => setTimeout(resolve, 0));
 
-    // Page Settings
-    class IndexPage extends React.Component {
-        render() {
-            return (<div>Hello, World</div>);
-        }
-    }
+test.beforeEach((t) => {
+  const routeMatcher = new RouteMatcher();
 
-    // Ready Sinon Spy
-    const spy = Sinon.spy(Router.prototype, 'renderer');
+  const historyManager = new HistoryManager();
+  historyManager.initialRouteMatcher(routeMatcher);
 
-    // The Router use RxJS to control async/await.
-    // So, First Mount is null.
-    const firstMountedActual = mount(
-        <Router history={history}>
-            <Route path="/" component={IndexPage}/>
-        </Router>
-    );
-    t.is(firstMountedActual.html(), null);
+  const request = new Request();
+  request.setHistoryManager(historyManager);
 
-    // renderer from componentWillMount
-    const component = await spy.getCall(0).returnValue;
-    const actual    = mount(component);
-    const expected  = mount(<IndexPage/>);
-    t.is(actual.html(), expected.html());
-
-    // renderer args Test
-    t.is(spy.getCall(0).args[0], '/');
-
-    // Restore Sinon Spy
-    Router.prototype.renderer.restore();
+  t.context.router = createRouter(historyManager, routeMatcher);
+  t.context.request = request;
 });
 
-Test.serial('match nest route', async (t) => {
-    // Location Settings
-    const history = createMemoryHistory();
-    history.push('/nest');
+test.serial("match single route", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/");
 
-    // Page Settings
-    class IndexPage extends React.Component {
-        render() {
-            return (<div>Hello, World</div>);
-        }
+  // Page Settings
+  class IndexPage extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
     }
+  }
 
-    class NestPage extends React.Component {
-        render() {
-            return (<div>Nest Page</div>);
-        }
-    }
+  // The Router use RxJS to control async/await.
+  // So, First Mount is null.
+  const mountedActual = mount(
+    <t.context.router history={history}>
+      <Route path="/" component={IndexPage} />
+    </t.context.router>
+  );
+  t.is(mountedActual.html(), null);
 
-    // Ready Sinon Spy
-    const spy = Sinon.spy(Router.prototype, 'renderer');
+  // wait to resolve promise.
+  await asyncFlush();
 
-    // The Router use RxJS to control async/await.
-    // So, First Mount is null.
-    const firstMountedActual = mount(
-        <Router history={history}>
-            <Route path="/" component={IndexPage}>
-                <Route path="/nest" component={NestPage}/>
-            </Route>
-        </Router>
-    );
-    t.is(firstMountedActual.html(), null);
-
-    // renderer from componentWillMount
-    const component = await spy.getCall(0).returnValue;
-    const actual    = mount(component);
-    const expected  = mount(<NestPage/>);
-    t.is(actual.html(), expected.html());
-
-    // renderer args Test
-    t.is(spy.getCall(0).args[0], '/nest');
-
-    // Restore Sinon Spy
-    Router.prototype.renderer.restore();
+  // renderer from componentWillMount
+  await mountedActual.update();
+  const expected = mount(<IndexPage />);
+  t.is(mountedActual.html(), expected.html());
 });
 
-Test.serial("default history", async (t) => {
-    const router = new Router({});
-    const history = router.getHistory();
-    t.is(history.location.pathname, "/");
+test.serial("match nest route", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/nest");
+
+  // Page Settings
+  class IndexPage extends React.Component {
+    render() {
+      return <div>Hello, World</div>;
+    }
+  }
+
+  class NestPage extends React.Component {
+    render() {
+      return <div>Nest Page</div>;
+    }
+  }
+
+  // The Router use RxJS to control async/await.
+  // So, First Mount is null.
+  const mountedActual = mount(
+    <t.context.router history={history}>
+      <Route path="/" component={IndexPage}>
+        <Route path="/nest" component={NestPage} />
+      </Route>
+    </t.context.router>
+  );
+  t.is(mountedActual.html(), null);
+
+  // wait to resolve promise.
+  await asyncFlush();
+
+  // renderer from componentWillMount
+  await mountedActual.update();
+  const expected = mount(<NestPage />);
+  t.is(mountedActual.html(), expected.html());
 });
 
-Test.serial('first rendering', async (t) => {
-    // Location Settings
-    const history = createMemoryHistory();
-    history.push('/');
+test.serial("unmatch route", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/nest");
 
-    // Page Settings
-    class IndexPage extends React.Component {
-        static initialPropsWillGet(props) {
-            t.is(props.pathname, '/');
-        }
+  // The Router use RxJS to control async/await.
+  // So, First Mount is null.
+  const mountedActual = mount(<t.context.router history={history} />);
+  t.is(mountedActual.html(), null);
 
-        static async getInitialProps(props) {
-            t.is(props.pathname, '/');
-            return {
-                message: 'world'
-            };
-        }
+  // wait to resolve promise.
+  await asyncFlush();
 
-        static initialPropsStoreHook(componentProps) {
-            t.is(componentProps.message, 'world');
-        }
-
-        static initialPropsDidGet(componentProps) {
-            t.is(componentProps.pathname, '/');
-        }
-
-        render() {
-            return (<div>Hello, {this.props.message}</div>);
-        }
-    }
-
-    // Ready Sinon Spy
-    const spy = Sinon.spy(Router.prototype, 'renderer');
-
-    // The Router use RxJS to control async/await.
-    // So, First Mount is null.
-    const firstMountedActual = mount(
-        <Router history={history}>
-            <Route path="/" component={IndexPage}/>
-        </Router>
-    );
-    t.is(firstMountedActual.html(), null);
-
-    // renderer from componentWillMount
-    const component = await spy.getCall(0).returnValue;
-    const actual    = mount(component);
-    const expected = mount(<IndexPage message="world"/>);
-    t.is(actual.html(), expected.html());
-
-    // renderer args Test
-    t.is(spy.getCall(0).args[0], '/');
-
-    t.plan(7);
-
-    // Restore Sinon Spy
-    Router.prototype.renderer.restore();
+  // renderer from componentWillMount
+  await mountedActual.update();
+  t.is(mountedActual.html(), null);
 });
 
-Test.serial('next rendering from Request `to`', async (t) => {
-    // Location Settings
-    const history = createMemoryHistory();
-    history.push('/');
-
-    // Page Settings
-    class IndexPage extends React.Component {
-        static async getInitialProps(props, prevProps) {
-            return {
-                message: 'world'
-            };
-        }
-
-        render() {
-            return (<div>Hello, {this.props.name}</div>);
-        }
+test.serial("default history", async (t) => {
+  // Page Settings
+  class IndexPage extends React.Component {
+    static initialPropsDidGet(componentProps) {
+      t.is(componentProps.pathname, "/");
     }
 
-    class UserPage extends React.Component {
-        static initialPropsWillGet(props, prevProps) {
-            t.is(props.pathname, '/user');
-            t.is(prevProps.pathname, '/');
-        }
-
-        static async getInitialProps(props, prevProps) {
-            t.is(props.pathname, '/user');
-            t.is(prevProps.pathname, '/');
-            return {
-                userName: 'oneut'
-            };
-        }
-
-        static initialPropsStoreHook(componentProps, prevComponentProps) {
-            t.is(componentProps.userName, 'oneut');
-            t.is(prevComponentProps.message, 'world');
-        }
-
-        static initialPropsDidGet(componentProps, prevComponentProps) {
-            t.is(componentProps.userName, 'oneut');
-            t.is(prevComponentProps.message, 'world');
-        }
-
-        render() {
-            return (<div>Hello, {this.props.userName}</div>);
-        }
+    render() {
+      return <div>Hello, World</div>;
     }
+  }
 
-    const spy = Sinon.spy(Router.prototype, 'renderer');
+  // The Router use RxJS to control async/await.
+  // So, First Mount is null.
+  const mountedActual = mount(
+    <t.context.router>
+      <Route path="/" component={IndexPage} />
+    </t.context.router>
+  );
+  t.is(mountedActual.html(), null);
 
-    // The Router use RxJS to control async/await.
-    // Test up to call renderer.
-    mount(
-        <Router history={history}>
-            <Route path="/" component={IndexPage}/>
-            <Route path="/user" component={UserPage}/>
-        </Router>
-    );
+  // wait to resolve promise.
+  await asyncFlush();
 
-    Request.to('/user');
+  // renderer from componentWillMount
+  await mountedActual.update();
+  const expected = mount(<IndexPage />);
+  t.is(mountedActual.html(), expected.html());
 
-    // renderer Test
-    const component = await spy.getCall(1).returnValue;
-    const actual    = mount(component);
-
-    const expected = mount(<UserPage userName="oneut"/>);
-    t.is(actual.html(), expected.html());
-
-    // renderer args Test
-    t.is(spy.getCall(1).args[0], '/user');
-
-    t.plan(10);
-
-    // Restore Sinon Spy
-    Router.prototype.renderer.restore();
+  t.plan(3);
 });
 
-Test.serial('Router props', async (t) => {
-    // Location Settings
-    const history = createMemoryHistory();
-    history.push('/');
+test.serial("first rendering", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/");
 
-    // Page Settings
-    class IndexPage extends React.Component {
-        render() {
-            return (
-                <div>
-                    <h1>Index</h1>
-                    <ul>{this.props.items.map((item, index) => (<li key={index}>{item}</li>))}</ul>
-                </div>
-            );
-        }
+  // Page Settings
+  class IndexPage extends React.Component {
+    static initialPropsWillGet(props) {
+      t.is(props.pathname, "/");
     }
 
-    const spy = Sinon.spy(Router.prototype, 'renderer');
+    static async getInitialProps(props) {
+      t.is(props.pathname, "/");
+      return {
+        message: "world"
+      };
+    }
 
-    // The Router use RxJS to control async/await.
-    // So, First Rendering is null.
-    const firstMoutedActual = mount(
-        <Router history={history} items={['foo', 'bar', 'baz']}>
-            <Route path="/" component={IndexPage}/>
-        </Router>
-    );
-    t.is(firstMoutedActual.html(), null);
+    static initialPropsDidGet(componentProps) {
+      t.is(componentProps.pathname, "/");
+    }
 
-    // renderer from componentWillMount
-    const component = await spy.getCall(0).returnValue;
-    const actual    = mount(component);
-    const expected = mount(<IndexPage items={['foo', 'bar', 'baz']}/>);
-    t.is(actual.html(), expected.html());
+    render() {
+      return <div>Hello, {this.props.message}</div>;
+    }
+  }
 
-    // renderer args Test
-    t.is(spy.getCall(0).args[0], '/');
+  // The Router use RxJS to control async/await.
+  // So, First Mount is null.
+  const mountedActual = mount(
+    <t.context.router history={history}>
+      <Route path="/" component={IndexPage} />
+    </t.context.router>
+  );
+  t.is(mountedActual.html(), null);
+
+  // wait to resolve promise.
+  await asyncFlush();
+
+  // renderer from componentWillMount
+  await mountedActual.update();
+  const expected = mount(<IndexPage message="world" />);
+  t.is(mountedActual.html(), expected.html());
+
+  t.plan(5);
+});
+
+test.serial("next rendering from Request `to`", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/");
+
+  // Page Settings
+  class IndexPage extends React.Component {
+    static async getInitialProps(props, prevProps) {
+      return {
+        message: "world"
+      };
+    }
+
+    render() {
+      return <div>Hello, {this.props.name}</div>;
+    }
+  }
+
+  class UserPage extends React.Component {
+    static initialPropsWillGet(props, prevProps) {
+      t.is(props.pathname, "/user");
+      t.is(prevProps.pathname, "/");
+    }
+
+    static async getInitialProps(props, prevProps) {
+      t.is(props.pathname, "/user");
+      t.is(prevProps.pathname, "/");
+      return {
+        userName: "oneut"
+      };
+    }
+
+    static initialPropsDidGet(componentProps, prevComponentProps) {
+      t.is(componentProps.userName, "oneut");
+      t.is(prevComponentProps.message, "world");
+    }
+
+    render() {
+      return <div>Hello, {this.props.userName}</div>;
+    }
+  }
+
+  // The Router use RxJS to control async/await.
+  // Test up to call renderer.
+  const mountedActual = mount(
+    <t.context.router history={history}>
+      <Route path="/" component={IndexPage} />
+      <Route path="/user" component={UserPage} />
+    </t.context.router>
+  );
+
+  // wait to resolve promise.
+  await asyncFlush();
+
+  // next rendering
+  t.context.request.to("/user");
+
+  // wait to resolve promise.
+  await asyncFlush();
+
+  // renderer update
+  await mountedActual.update();
+  const expected = mount(<UserPage userName="oneut" />);
+  t.is(mountedActual.html(), expected.html());
+
+  t.plan(7);
+});
+
+test.serial("Router props", async (t) => {
+  // Location Settings
+  const history = createMemoryHistory();
+  history.push("/");
+
+  // Page Settings
+  class IndexPage extends React.Component {
+    render() {
+      return (
+        <div>
+          <h1>Index</h1>
+          <ul>
+            {this.props.items.map((item, index) => (
+              <li key={index}>{item}</li>
+            ))}
+          </ul>
+        </div>
+      );
+    }
+  }
+
+  // The Router use RxJS to control async/await.
+  // So, First Rendering is null.
+  const mountedActual = mount(
+    <t.context.router history={history} items={["foo", "bar", "baz"]}>
+      <Route path="/" component={IndexPage} />
+    </t.context.router>
+  );
+  t.is(mountedActual.html(), null);
+
+  // wait to resolve promise.
+  await asyncFlush();
+
+  // renderer from componentWillMount
+  mountedActual.update();
+  const expected = mount(<IndexPage items={["foo", "bar", "baz"]} />);
+  t.is(mountedActual.html(), expected.html());
 });
